pyproject.toml
[tool.poetry]
name = "weaviate-starter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
weaviate-client = "^3.25.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
